[update] シミュレーションの細かいバグ修正と仕様変更
- numpyが1.16とかだと動かない問題を修正
- ジョイスティックのXボタンを押した瞬間からシミュレーションがスタートするように変更
- clockを落として負荷を少なく
-
情報信頼度が落ちるスピードが小さいと、エージェントが止まってしまう問題があったのでデフォルト値を上昇させる
- 普通の持続的被覆のゲインのデフォルト値を上昇させる

+++ script/virtual_drone.py
@@ -119,7 +119,10 @@ class RigidBodyMotion(object):
            self.publish_pose()
            self.velocity_pub.publish(self.velocity)
            self.tf_broadcat()
-           self.g = g_reprojection(self.g)
+           # if numpy == 1.13.1
+           # self.g = g_reprojection(self.g)
+           # otherwise
+           self.g = g_reprojection(self.g.astype(np.float64))
            self.rate.sleep()
            # print self.g
         return 0
